services/astro-engine/python/astro_calculation_engine/pyproject.toml
[project]
name = "astro-calculation-engine"
version = "0.1.0"
description = "Swiss-Ephemeris based astrology calculation engine with astro_package-compatible midpoint exports"
requires-python = ">=3.10"
dependencies = [
  "pyswisseph>=2.10.3.2",
  "timezonefinder>=6.5.0",
  "python-dateutil>=2.9.0",
]

[project.scripts]
astro-engine = "astro_engine.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
